+++ package.json
@@ -36,6 +36,7 @@
     "react-router-dom": "5.2.0",
     "redux": "4.0.5",
     "redux-devtools-extension": "2.13.8",
-    "redux-thunk": "2.3.0"
+    "redux-thunk": "2.3.0",
+    "reselect": "^4.0.0"
   }
 }
