+++ src/components/movies-list/movies-list.jsx
@@ -9,7 +9,6 @@ import {loadMovieList} from '../../api/api-actions';
 import {ActionCreator} from '../../store/action';
 import Loading from '../loading/loading';
 import NothingToDisplay from '../nothing-to-display/nothing-to-display';
-import {getMovies, getGenre} from '../../store/data/selectors';
 
 const getMovieCardComponent = (id, rest, onMovieCardMouseEnter, onMovieCardMouseLeave, currentMovieId) => {
   return <MovieCard key={id} movieId={id} {...rest} onMovieCardMouseEnter={onMovieCardMouseEnter} onMovieCardMouseLeave={onMovieCardMouseLeave} currentMovieId={currentMovieId} />;
@@ -66,10 +65,7 @@ MoviesList.propTypes = {
 };
 
 const mapStateToProps = (state, componentProps) => {
-  const movies = getMovies(state);
-  const genre = getGenre(state);
-
-  const targetMovies = getMovieByPathName(movies, genre, componentProps);
+  const targetMovies = getMovieByPathName(state, componentProps);
 
   return {
     targetMovies
