+++ src/store/data/selectors.js
@@ -1,5 +1,6 @@
+import {createSelector} from 'reselect';
 import {NameSpace} from '../root-reducer';
-
+import {DEFAULT_GENRE, AMOUNT_OF_SIMILAR_MOVIES} from '../../../src/utils/constatns';
 
 export const getMovies = (state) => state[NameSpace.DATA].movies;
 
@@ -10,3 +11,35 @@ export const getSelectedMovieReviews = (state) => state[NameSpace.DATA].selected
 export const getGenre = (state) => state[NameSpace.DATA].genre;
 
 export const getIsPromoLoadedFlag = (state) => state[NameSpace.DATA].isPromoLoaded;
+
+export const getFavoriteMoviesSelector = createSelector(
+    getMovies,
+    (movies) => movies.filter((movie) => movie.is_favorite)
+);
+
+export const getMoviesByGenreSelector = createSelector(
+    getMovies,
+    getGenre,
+    (movies, selectedGenre) => {
+      if (selectedGenre !== DEFAULT_GENRE) {
+        return movies.filter(({genre}) => genre === selectedGenre);
+      }
+
+      return movies;
+    }
+);
+
+export const getSimilarMoviesSelector = createSelector(
+    getMovies,
+    getSelectedMovie,
+    (movies, movie) => {
+      if (movie) {
+        const {id: clickedMovieId, genre: similarGenre} = movie;
+        const allMoviesSameGenre = movies.filter(({id, genre}) => genre === similarGenre && id !== clickedMovieId);
+        const similarMoviesToShow = allMoviesSameGenre.slice(0, AMOUNT_OF_SIMILAR_MOVIES);
+        return similarMoviesToShow;
+      }
+
+      return undefined;
+    }
+);
